move test to works
