+++ src/app/navbar.js
@@ -61,7 +61,6 @@ export default function Navbar() {
         ["關於我們", "/about"],
         ["歷年作品", "/works"],
         ["Q & A", "/qna"],
-        ["Test", "/test"]
     ];
 
     function toggleNav() {
